Moved UI code into pages and renamed prompt.helper.ts to prompt.controller.ts
